+++ main.cpp
@@ -1,9 +1,11 @@
 #include <motmotCore.hpp>
 
-#include <QCoreApplication>
+#include <signal.h>
+
+#include <QApplication>
 
 int main(int argc, char *argv[]) {
-    
+       
     if (argc < 5) {
         std::cout << std::endl
             << "Usage: ./MotMot path_to_vocabulary path_to_settings if_visualize(true or false) imu_port"
@@ -11,9 +13,18 @@ int main(int argc, char *argv[]) {
         return 0;
     }
     
-    QCoreApplication motmotApp(argc, argv);
+    QApplication motmotApp(argc, argv);
 
-    motmotCore core(nullptr, argc, argv);
+    motmotCore core(QCoreApplication::instance(), argc, argv);
+    
+    struct sigaction hup;
+    hup.sa_handler = core.callSignalHandler;
+    sigemptyset(&hup.sa_mask);
+    hup.sa_flags = 0;
+    hup.sa_flags |= SA_RESTART;
+    if (sigaction(SIGINT, &hup, 0))
+        return 1;
 
-    return motmotApp.exec();
+    int retVal = QCoreApplication::exec();
+    return retVal;
 }