+++ ChangeLog.md
@@ -1,3 +1,8 @@
+# v0.0.2
+
+- Improvement in application architecture
+- Fixed issues in camera functionality
+
 # v0.0.1
 
 - Read IMU data from serial port